+++ app/auth/oauth2.py
@@ -1,9 +1,9 @@
+from typing import List
 from uuid import UUID
 
 from app.auth.jwt_token import decode_token
 from app.core.exceptions import ForbiddenException, Unauthorized
-from app.routers.auth.schemas import User
-from app.routers.workplace import Role, UserAssignedWorkplace
+from app.routers.auth.schemas import Role, User, UserAssignedWorkplace
 from fastapi import Depends, Path
 from fastapi.security import HTTPBearer
 
@@ -19,7 +19,7 @@ async def get_current_user(token: str = Depends(oauth2_scheme)) -> str:
 
 
 class RoleChecker:
-    def __init__(self, role: list[Role]):
+    def __init__(self, role: List[Role]):
         self.role = role
 
     async def __call__(self, user: User = Depends(get_current_user), workplace_id: UUID = Path(...)) -> str:
@@ -30,7 +30,7 @@ class RoleChecker:
         )
         if workplace is not None and workplace.role in self.role:
             return user
-        raise ForbiddenException("Не авторизован.")
+        raise ForbiddenException("Нет прав.")
 
 
 # доступно только админу
