+++ app/routers/__init__.py
@@ -1,8 +1,7 @@
-from app.routers.auth import User
+from app.routers.auth import User, UserAssignedWorkplace
 from app.routers.auth import router as register_router
 from app.routers.sprint import Sprint
 from app.routers.sprint import router as sprint_router
-from app.routers.workplace import UserAssignedWorkplace
 
 list_of_routes = [
     register_router,
