+++ app/routers/auth/__init__.py
@@ -1,4 +1,4 @@
 from .router import router
-from .schemas import User
+from .schemas import User, UserAssignedWorkplace
 
-__all__ = ["User", "router"]
+__all__ = ["User", "UserAssignedWorkplace", "router"]
