--- app/routers/workplace/__init__.py
@@ -1,3 +1,0 @@
-from .schemas import Role, UserAssignedWorkplace
-
-__all__ = ["UserAssignedWorkplace", "Role", "router"]
